Rename Java packages. Add dependencies for DynamoDB.
