--- src/main/java/com/aws/codestar/projecttemplates/EnaTsApplication.java
@@ -1,13 +1,0 @@
-package com.aws.codestar.projecttemplates;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EnaTsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EnaTsApplication.class, args);
-    }
-
-}
