--- src/main/java/com/aws/codestar/projecttemplates/controller/UploadController.java
@@ -1,12 +1,0 @@
-package com.aws.codestar.projecttemplates.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class UploadController {
-    @GetMapping("/upload")
-    public String uploadForm() {
-        return "upload";
-    }
-}
