--- src/main/java/com/aws/codestar/projecttemplates/json/JsonMapped.java
@@ -1,23 +1,0 @@
-package com.aws.codestar.projecttemplates.json;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-
-public abstract class JsonMapped<T> {
-    ObjectMapper mapper = createMapper();
-
-    public String toJson() throws JsonProcessingException {
-        return mapper.writeValueAsString(this);
-    }
-
-    public T parse(String json) throws JsonProcessingException {
-        return mapper.readValue(json, (Class<T>) this.getClass());
-    }
-
-    private ObjectMapper createMapper() {
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.registerModule(new JavaTimeModule());
-        return mapper;
-    }
-}
