+++ src/main/java/com/ena/timesheet/EnaTsApplication.java
@@ -1,0 +1,13 @@
+package com.ena.timesheet;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EnaTsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EnaTsApplication.class, args);
+    }
+
+}
