+++ src/main/java/com/ena/timesheet/controller/UploadController.java
@@ -1,0 +1,12 @@
+package com.ena.timesheet.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UploadController {
+    @GetMapping("/upload")
+    public String uploadForm() {
+        return "upload";
+    }
+}
