+++ src/main/java/com/ena/timesheet/phd/PhdTemplate.java
@@ -1,0 +1,19 @@
+package com.ena.timesheet.phd;
+
+import java.util.List;
+
+public class PhdTemplate {
+    public PhdTemplate(List<PhdTemplateEntry> entries) {
+        this.entries = entries;
+    }
+
+    public List<PhdTemplateEntry> getEntries() {
+        return entries;
+    }
+
+    public void setEntries(List<PhdTemplateEntry> entries) {
+        this.entries = entries;
+    }
+
+    private List<PhdTemplateEntry> entries;
+}
